adjustments were made to the withSafeArea HOC
The react-native-paper package already handles safe areas in components
such as App Bar, Bottom Bar and Drawer. In order to ensure that the
content of the screen also has this guarantee, we extended the HOC
arguments to define which boundaries we want to add. By default we only
add it to the bottom, because all screens end up having the App Bar.