wip: delete account action verification
- the DeleteAccountAction component was created, which visually only has
one button, but in this component we have a handler that checks whether
or not the user can delete the account;
- the getCoolingUnitsEmployees and getCoolingUnitsByStatus methods were
created;
- the getCoolingUnitsEmployees method is responsible for returning the
registered employees associated with a cooling unit. We use this method
to check whether the user trying to delete the account is the last user
of the cooling unit;
- the getCoolingUnitsByStatus method helps to check whether a user
(whether farmer/cooling user or not) or a company has cooling units with
a status of empty and not empty.