sign up methods were added through the invitation
- methods were added to the AuthService that deal with the mutation of
account creation through the invitation;
- adjustments were made to the Invite component to make these same
mutations according to the type of user;
- a parameter called accountProfile was added to the Auth browser, which
allows defining the type of account in the default values of the form
state. This is so that as soon as the sign-up made through the
invitation is successful, the user is redirected to the SignIn screen
and leaves the group pre-populated properly.