chore: http requests made in notifications have been reduced
- changes were made to the processing of notifications to be done in two
parts, one to obtain the original datums and then to process the
messages with the translations. The last step is to add some context to
each of the notifications to ensure that we don't need to make
subsequent requests to obtain this same data;
- changes were made to the component architecture to divide
responsibilities between the parent and the child components;
- data such as farmers, cooling units and crops were preloaded to ensure
that we do not have any delays related to the resolution of these
promises.